Separate client conveniences. Add JSON typealias & Header struct.
